feat: refactored CLI into it's own file, allows for cleaner importing for testing
